requirements.txt
colorama==0.4.6
gitdb==4.0.11
GitPython==3.1.43
smmap==5.0.1
tqdm==4.66.5
